New feature! Standalone Arduino FVM with keyboard and display.
The fvm.c and tape-clcd.ino in this commit are correctly configured to run the ts.fl tape server on an Arduino Mega 2560 directly connected to a 20x4 character LCD and a PS/2 keyboard. The user can interact with the tape as usual. The big difference between this and previous commits is that all this is self-contained on one Arduino board (rather than using separate boards/computers, connected by a serial connection, for the tape server and the tape terminal).