Initial publication of pre-alpha version of tape-gameduino.ino
This is an early pre-alpha implementation of a tape terminal for Arduino Uno using a MOD-VGA-32MB shield (a Gameduino 1 clone) to drive a VGA monitor for monochrome text display at 40 colums x 16 rows. The versions of 'tape-gameduino.ino', 'fvm.c' and 'ts.fl' in this commit are correctly configured for testing this tape terminal via serial connection to an FVM instance running on Linux but without using split-tape mode. This implementation has some incorrect behaviour which is yet to be fixed, as noted in the 'tape-gameduino.ino' source, but is somewhat usable.