New Feature! Added support for Arduino Uno
The versions of 'fvm.c', 'ts.fl' and 'tape.c' in this commit are correctly configured for running a tape server on an FVM instance on an Arduino Uno connected by serial connection over USB to a tape terminal running on a Linux computer. This commit contains fixes to 'fvm.c' such that tracing is now working on Arduino Uno, since bugs in PROGMEM usage have been fixed. RAM usage is now acceptable.