Arduino Mega now runs fvmtest almost correctly
Fixed a number of silent bugs only discovered after running the whole fvmtest suite for the first time (in 4 parts due to 'rom.h' array index size limit on this platform) on the Arduino Mega. Everything now passes except that the total number of restarts is 238 not the expected 241 (which suggests some remaining hidden failures).