Added experimental binary incbin support
The experimental FVMO_INCORPORATE_BIN option causes a 'rom.fp' bytecode binary file to be incorporated into the FVM executable rather than doing so via a C array in 'rom.h'. Although it works on an Arduino Mega, it only works for programs less than 32 kiB in size anyway, so it is not a solution to the 32 kiB barrier.