Finished work on .incbin. Abandoned attempt to break 32 kB barrier.
The FVMO_INCORPORATE_BIN option (binary 'rom.fp' include) is now working nicely on Arduino Mega but remains limited to 32 kB on 8-bit Arduinos, the same limit as FVMO_INCORPORATE_ROM alone (text 'rom.h' include). It's not worth spending more time on trying to break this barrier as there are too many complicating factors depending on the hardware of the particular board used and it is frustratingly difficult to quickly find the necessary documentation. Therefore, sadly, this is not a solution for how to run the whole 80 kB fvmtest suite in one go.