Verified that Arduino Mega 2560 really did pass entire fvmtest suite
Did so by running the fvmtest suite on Freetronics EtherDue divided into the exact same 4 parts as were run on the Arduino Mega 2560 and found the actual number of restarts to be in every case identical. Thus the reason why total actual restarts was 238 compared to reported expected restarts of 241 is simply that splitting the test suite run into n parts will result in n-1 fewer actual restarts since for each run after the first, 1 restart is replaced by the manual start of the run.