Withdrawing support for Arduino Uno and TI Launchpads
Henceforth decided to target (at least for now) only microcontrollers having at least 128 kB of flash and at least 8 kB of RAM, for reasons of project productivity. The entry-level Arduino board for running Freeputer is thus the Arduino Mega 2560. It works fine and is easy to support. See extensive comment in 'fvm.c' for details.