Milestone! First time flc has compiled itself on a microcontroller!
Although the self-hosted Freelang compiler 'flc.fl' has always been able to compile itself when running on Linux in a large FVM instance having 16 MiB of RAM (in about 1 second on a 3 GHz i5 box), this is the first time ever that it has compiled itself on a microcontroller. This feat was achieved on a chipKIT Max32 board, using the FVMC_CHIPKIT_SD53_FLC_SMALL configuration, hosting an FVM instance with only 4 kiB of RAM (in about 4.5 hours, running on the 80 MHz PIC32; the extremely slow speed was due to an SD card, using only the regular Arduino SD Library, acting as stdblk and being used instead of RAM for dynamic storage during compilation; much faster speed would be possible with faster storage media). The init and compile (first pass) steps only took half an hour; it was population of references in the pmodel from the ptab (second pass) that took about 4 hours. The source code was read from 'std.in' on the SD card and output sent over serial connection to a Linux box for analysis. This is, of course, impractically slow but nevertheless is a huge milestone as it clearly demonstrates that no operating system is required for the 'flc.fl' Freelang compiler to compile itself. That is freedom in software development: knowing you can recompile the compiler without needing any operating system to recompile it on (and indeed on a mere microcontroller, not even requiring a computer proper). That is a future-proof solution. What should follow next is a simple hardware implementation of external RAM in order to solve the performance problem; but that can wait for now as the software project proceeds.