Raspberry Pi U-Boot: first successful FVM run
The 'fvm.c' in this commit was used to run an FVM instance bare-metal on the Raspberry Pi as a standalone U-Boot application for the first time. It run a tiny Freelang program printing a few characters to stdout. Of course, the FVM has always run on Linux on Raspberry Pi without a problem but this is the first time that Freeputer has been successfully deployed to U-Boot and run without any operating system on the Pi; this paves the way to run the FVM in future, bare-metal, on many boards to which U-Boot has been ported. This is a very early, unstable, incomplete FVM implementation for U-Boot (where so far it supports only stdin and stdout; not yet stdblk, stdimp or stdexp).