Withdrawing support for bare-metal Raspberry Pi
The 'fvm.c' in the newly added archival directory 'pre-alpha/pre-alpha1.1/retired/fvm/c/src/pi-bare/' can successfully run the 'flc.fl' compiler on an FVM instance running bare-metal on the Raspberry Pi (using only U-Boot and no operating system) so as to compile itself. However, it has become clear that the complexity (and partially closed nature) of such hardware is inimical to freedom. That is, the Raspberry Pi and similar ARM Linux boards are very nice for conveniently running Linux but do not provide broader freedoms such as the freedom to easily be used without an operating system.