Ongoing enhancement of 'fvmtest.fl' for FVM Lite
On Linux, an FVM Lite instance now passes the XT_S1_FVMLITE subset of the 'fvmtest.fl' test suite which is conveniently generated by the new 'xt.sh' script. The latter will soon be enhanced to generate all subsets of the test suite, so as to allow them to subsequently be easily compiled and run one at a time on an FVM Lite instance to test the function of that instance. This will be particularly useful for testing the FVM on microcontrollers but also for testing any FVM Lite instance anywhere. Of course, running the 'fvmtest.fl' test suite is straightforward for FVM Heavy: the entire test suite is simply run at once. This approach is a compromise so as to keep the whole suite in a single file with no code duplication.