Improved strategy of 'xt.sh' for fvmtest
Previous strategy was ineffective. New strategy works by automatically uncommenting only the section to be run, after first commenting-out all subsections en masse. This has proven to be effective and is working for the first three subsections of 'fvmtest.fl'. All that remains is to roll this out for all subsections.