'xt.sh' with 'fvmtest.fl' tests FVM Lite in 25 subsections
The tedious task of breaking up 'fvmtest.fl' into 25 subsections, to be done automatically by running 'xt.sh', has been successfully carried out. Manually running each subsection on an FVM Lite instance on Linux demonstrated that the complete suite so passes (although it is not immediately clear if the total of 223 restarts is good or not... it probably is either correct or close to correct). The benefit of this work is that FVM Lite instances can now be (somewhat tediously but effectively) tested on any platform, including microcontrollers, without any cludges (such as using artificially large ROM sizes).