Updated flc compiler to 1.0.0.2 for FVM 1.0
This update fixes another bug found in the flc compiler versions 1.0.0.0 and 1.0.0.1 for FVM 1.0. This update is provided as a courtesy to users of Freeputer 1.0. However, this update HAS NOT BEEN EXTENSIVELY TESTED since development effort is now focused on Freeputer 1.1 not Freeputer 1.0. See comments in 'flc.fl' for further details.