Updated flc compiler to 1.0.0.3 for FVM 1.0
This update demonstrates how to increase compiler capacity so larger Freelang programs can be compiled by flc. This update is provided as a courtesy to users of Freeputer 1.0. However, this update HAS NOT BEEN EXTENSIVELY TESTED since development effort is now focused on Freeputer 1.1 not Freeputer 1.0. See comments in 'flc.fl' for further details.