Standard grid size reduced to 40 rows x 12 columns.
After careful consideration, it was decided that the previous size of 40 rows x 15 columns was excessive for a standard minimal requirement, particularly in the light of the need to be easily portable to small devices, simple hardware and mobile devices. The new size very conveniently requires less than half a kilobyte of RAM to represent the state of a minimal 7-bit ASCII monochrome standard grid and very conveniently is exactly a quarter the size of a standard 80x24 terminal emulator.