Standard grid size reduced to 40 columns x 12 rows.
The purpose of this commit, which contains only an updated timestamp on the README files in the grid directory, is to explicitly correct a typo in the previous commit message in which the words 'rows' and 'columns' were accidentally swapped. The standard grid is of course now 40 columns (not rows) wide and 12 rows (not columns) deep.