Freeputer Assembler: new concept
With this commit the implementation of the new FW32 (fixed-width 32-bit instructions) strategy for FVM 2.0, as detailed in the last commit of README.md, has begun. In the course of this work it has become apparent that the existing flx.rb and flc.fl compilers are not especially well suited to the present effort; essentially they are more complicated than need be for the newly simplified instruction set. It might be better to write a simple assembler for FVM 2.0 bytecode. Therefore this commit introduces the concept of a Freeputer Assembler for FVM 2.0 bytecode. The assembler is intended to be much easier to implement than was a Freelang compiler. The prototype assembler will be written in JavaScript.