Added push and pop support.
Note: the immediate defines the address of the pointer for these stack operations. The immediate does NOT define the thing to be pushed or popped, since that is always the sole register of the single register machine. In other words, these push and pop instructions allow any cell in memory to be used as a stack pointer.