Parking srm.s for now. Very promising.
Experiments have confirmed that this is a very promising strategy. However, this outcome suggests that perhaps it is possible to go another order of magnitude simpler and still get good results. Therefore parking srm.s for now and switching instead to tvm.s Tiny Virtual Machine experiments which will follow a similar but smaller pattern.