Start of 'miscvm1.s' MISC experiment.
Explanation: Although 'tvm.s' in general looks very promising and is currently the front-runner for an FVM 2.0 foundation, it still seems heavier than need be and not as trivially easy to port as would be ideal. Also I am not convinced that it would be sufficiently easy to implement in hardware. Therefore parking 'tvm.s' for now while attempt to find an even more minimal solution.