DECISION: 'tvm.s' is again the front-runner.
Experiments with more registers such as 'lm.s', '8rm.s' and 'miscvm1.s' turned out to be harder to grok than 'tvm.s' and introduced more difficult problems of comprehension. Much of this is about finding a solution which is easy for humans to understand. Modern software engineering has become so absurdly complex and so dependent on compilers (which no single human being could ever fully understand) that the instruction sets of modern CPUs are now largely incomprehensible. Freeputer, in contrast, aims to have an instruction set easily understood by humans. In FVM 1.0 this was achieved by a stack machine but that choice brought with it considerable complexity of VM implementation.