Adding C implementation as a sanity check.
So far the 'tvm.s' looks extremely promising. However, it is written in x86 assembly language. Before going too much further it is now worth checking that this approach also looks good when implemented in C, since there will not always be time to hand-roll implementations in assembly language for various platforms.