REJECTED all current approaches. Decided to go MISC.
The reason is that a deep dive into how to eliminate all undefined behaviour in the C implementation shows that doing so is diabolically difficult and exponentially more so the more lines of code there are in the VM implementation. This is a dead end. FVM 1.0 already has provided a reasonably undefined-behaviour-free solution and as a result is not trivial to implement. If we really want QUALITY and EASE OF IMPLEMENTATION and SIMPLICITY then MISC is the only way to go because anything else will require too many lines of code to: (a) be easy to ensure there is no undefined behaviour in; and (b) be nicely easy to implement in assembly language without the horrors of trying to understand the behaviour of millions of lines of C-compiler code.

Doing this requires LETTING GO OF PERFORMANCE and STOPPING ALL PREMATURE OPTIMIZATION. Forget SPEED. Go for QUALITY and FREEDOM instead.

In other words: C is the problem not the solution.