MISC: decided to use static jumps only.
This is to preserve independence of programs from whatever the underlying architecture and structure of the program space of the native platform happens to be. To have dynamic jumps we would have to enforce instruction width there, which would lead to forcing the use of bytecode interpretation thus precluding native implementation, which is not acceptable at this time.