MISC: Eased comprehension by humans.
Did so by restricting literals to 24 bits, and adding msb instructions to handle the most significant byte, so as to allow convenient instructions such as 'immd 100' rather than requiring 'imm 100' followed by 'immd'. This is a trade-off since some literals will require two instructions to set up, which will be harder to understand, but the most commonly used literals do not require this.