MISC: Resumed use of vI!
This was unexpected. However, after successfully eliminating vS, vD and vP it became apparent that resuming the use of vI, although it results in more verbose syntax, greatly reduces the number of opcodes required; it also has the ancilliary benefit of allowing the use of nice 31-bit literals augmented by the Set() opcode. This approach brings the core of the parent VM implementation down below 100 lines of code. This is the kind of implementation simplicity we are looking for and is worth the more verbose syntax.