MISC: Further rationalization.
Note: this VM aims to be so simple that it is easy to implement without accidentally including undefined behaviour in its implementation even when that implementation is written in C. Its design eliminates anything (such as multiplication and division) likely to make achieving that difficult. Thus all integers are unsigned. Therefore to avoid confusion, and to avoid assumptions which may be incorrect in some contexts, the VM intentionally does not provide any less-than or greater-than operators. Instead all conditionals are in essence bitwise.