MISC: Child 15x slower than parent.
The child VM, virtualized inside the parent VM, executes an empty rpt loop about 15 times slower than the parent. Note that asm("") has to be used in the parent rpt definition to prevent gcc optimizing away the empty loop, and that this is NECESSARY to achieve the same behaviour in the parent as in the child and to prevent control being taken away from the programmer and given to the compiler.