MISC: Performance: 12.9 secs 0x7fffffff nop repeats.
This 'qmisc.c' uses crude optimization of instruction decoding simply by splitting into simple and complex instructions. Can easily be further optimized. Will do so later. It is IMPORTANT to note that this compares to 17.6 seconds by 'qmisc-interpreted-abandoned.c' which was a typical C implementation of a bytecode VM using a switch statement! That is, even with -O3, gcc produced code which was about 36% slower when running as the parent VM than is this virtualized child VM rolled by hand. Therefore do not be tempted to return to an approach similar to 'qmisc-interpreted-abandoned.c'. This current approach is superior. Considering that the goal is fast virtualization of a child within a parent, this current approach is an order of magnitude faster.