MISC: Calling this 'qmisc.c' 0.0.1.0. Good design.
At this point we have the basic design more or less in place. It can be tweaked a bit but this is basically it. What is missing is I/O. Performance of the parent native VM is excellent and performance of the child VM (virtualized within the parent) is surprisingly good. The child does 0x7fffffff empty repeats (nop, rpt) in about 11 seconds. The parent does them (suspiciously) in about a second or less. Main limitation of the design is maximum 1 GB data memory. Simplicity here is amazing, the core of the VM implementation is only about 75 lines of C (from line 56 to line 131). IMPORTANT LESSON LEARNED: Comparing this 'qmisc.c' to the FVM 1.0 'fvm.c' shows this core is an order of magnitude simpler and that this is definitely the right approach if one wishes to be reasonably able to eliminate all undefined behaviour. In retrospect, it is too hard to be absolutely sure of that with a C implementation (as opposed to an assembly-language implementation) of FVM 1.0. With C, smaller and simpler is better. Note: this 'qmisc.c' contains some bugs; in particular it is not yet masking and translating child program addresses and child data memory addresses, so its true performance will be a bit slower than this.