MISC: Start of 'qmisc.s' x86 assembly language implementation.
Now that the initial proof of concept 'qmisc.c' has been shown to work in C, it is time to begin an implementation in assembly language. Freeputer is intended as a platform which does not necessarily require a C compiler to create a native VM implementation. Not only does this provide the ultimate in portability, even to platforms for which no C compiler yet exists, but it also avoids undefined behaviour which the use of C might accidentally introduce.