MISC performance: x86=30s, m32=19s, m64=11s.
Unfortunately the child VM of 'qmisc.s' takes 30 sec to do 0x7fffffff nop repeats, compared to 11 sec for 64-bit-compiled 'qmisc.c' and 19 sec for 32-bit compiled 'qmisc.c'. However, those last two use impenetrable gcc fu: good luck trying to disassemble those executables and understand them. This impenetrable fu is contrary to the spirit of Freeputer which is supposed to be lightweight, easy to understand, and not dependent on millions of lines of gcc code which no single human can ever hope to understand. In any case, it should be borne in mind that here we are comparing the performance of child VMs (self-virtualization within the parent) not the parent VMs; seen in that context, this performance is not unreasonably slow. Also, the parent VMs display the opposite behaviour: the 'qmisc.s' parent does 0x7fffffff nop repeats in 1.4 secs whereas the 'qmisc.c' parent takes 4 secs regardless of whether it is compiled for 32-bit or 64-bit. Thus, in summary, the x86 assembly-language implementation is about 3x faster as a parent and up to 3x slower as a child. It seems likely therefore that the best compromise might be to run on an x86 assembly-language implementation which interpreted bytecode (best of both worlds: reasonable performance and reasonable simplicity).