MISC x86: *Hacked* 'qmisc.s' into an interpreter.
This is a quick and ugly experiment. Performance of the parent implemented as a bytecode interpreter was 17 seconds for 0x7fffffff nop repeats, which is comparable to the child VM of 'qmisc.c' on -m32 but about twice as fast as the previous self-virtualized 'qmisc.s' child. The lesson learned from this is: IT IS NOT WORTH IT. That is, implementing the parent as a bytecode interpreter (without relying upon C to do so, already found to be slower than self-virtualization in C and therefore pointless) is complex and painful because to do so one must learn the details of the target platform. Painful. Delightfully easy is simply to use a self-virtualized child; the only problem with that is performance, which although good on C is much worse on asm. The correct approach is to keep optimizing the self-virtualization until performance comes up to an acceptable level, rather than investing time and energy in native bytecode implementations.