MISC x86: Optimization: using ld x86-32, 4.2 sec.
This is the fastest combination known so far. All future optimization (that is, improving the algorithm for self-virtualization) will be done on this base until proven otherwise. This figure is for 0x10000000 nop repeats by the child VM (i.e. using self-virtualization). The native parent VM does this in 0.26 seconds. That is, using 'qmisc.s' the child is about 16 times slower than its parent. The ratio is smaller on C but that is not directly relevant here.