MISC: Parking as 'qmisc-promising.c'.
Explanation: Recent uncommitted experiments with porting 'qmisc.s' from x86 to ARM hit some difficult technical problems which essentially highlighted the frustratingly closed nature of ARM as a platform, its dependency on complex compilers, and the weaknesses of open-source development tools for ARM assembly language. The problems were mainly ones of documentation. The port is not impossible; it's just extremely annoying to the point where one does not wish to bother. Rewriting from scratch would be easier (but still annoying). The bottom line is that, in effect, x86 (although a closed architecture) is a more open platform and easier to use with open-source tools. Considering these problems again highlighted the need to maximize hardware freedom by ensuring that the VM does not require more RAM than is typically available on microcontrollers, thus forgoing the need for an operating system and forgoing the need to use such problematic CPU architures. Much better is to head in the direction of using either microcontrollers or open-source soft CPUs implemented in FPGAs. Either way, excessive RAM dependency not only leads to monolithic (rather than modular) design but also drastically limits hardware freedom. After much thought, a decision was reached that 64 kB is the maximum amount of RAM which is compatible with both modular design and hardware freedom. Therefore 'qmisc.s' is about to be simplified and rewritten to allow not more than 64 kB of RAM. It may be better, for similar and corresponding reasons, to move from a 32-bit architecture to a 16-bit architecture (while not forgoing the ability to address large storage via I/O). Lastly, this work also made apparent that further simplification would be advantageous.