MISC: Parking as 'qmisc-native.c'.
Very sadly, am about to abandon native implementation because it is not reliably portable because gcc local labels are not reliably portable; for example, they don't seem to work on Arduino. Thus although a native implementation works fine in assembly language it is not portable in C. This is all because of the link instruction and vL. The only simple, reliable way forward is to go to a bytecode interpreter for the parent (which was previously found to be both slower and, in some respects, more complex). Sigh...