MILESTONE DECISIONS: XS, S, M, L, XL.
After a great deal of thought over several weeks, finally settled on a scheme of 5 standard sizes and accepted that the VM shall trap upon illegal instruction or memory access out of bounds, since a program which does those things is essentially meaningless (or unpredictable) and therefore should stop running. There is no such thing as a perfect VM design for all situations and all purposes; the design decisions are always a trade-off. However, this appears to be a very reasonable compromise which will support great hardware freedom, quality, and modular virtualization of algorithms in a manner which does not necessarily require an operating system.