About to experiment with true MISC.
A decision needs to be made whether to commit to a truly MISC approach or to continue favouring performance over simplicity. Code smells such as the COPY instruction seem to have crept in, bloating the instruction set. Yet design thinking has repeatedly concluded that the entire address space should be RAM, which means that self-modifying code is entirely possible and reasonable here, which means that a truly MISC instruction set could work. On the other hand, performance would be pretty poor. But perhaps that is of lesser importance overall. Parking 'fvm2.js' and 'fvm2-full.js' in preparation for converting 'fvm2.js' to a truly MISC approach as an experiment.