MISC: Enhanced assembler.
Note: 'prg.c' is just a copy of most of the content of 'prg.js' but renamed so that it can be viewed with C syntax highlighting. The syntax of the assembly language has deliberately been chosen to be compatible with C syntax (in the right context, see 'qmisc-native.c') so that it can be compiled to native code and so that it can be used with nice syntax highlighting for free. To do this the VM first needs to be changed back to using Harvard architecture and that will be done soon.