MISC: Ready to begin design patterns.
Primarily now working with 'fvm2.js' in the first instance (yet to bring 'qmisc.c' and 'qmisc.s' up to date). Things have now progressed sufficiently to begin experimenting in earnest with design patterns of writing programs in the assembly language. This will be done in 'prg.js' which shall be mirrored in 'prg.c' (the latter for native compilation, the former for running in the JavaScript interpreter). One more small task to be done soon is to convert the implementation back to Harvard architecture as that is better suited to this twin strategy.