MILESTONE: VICTORY FOR STACK MACHINES!
The thing is to be as simple as possible but no simpler. Considering the difficulty for human beings of writing programs and factoring programs, stack machines are the superior choice for freedom because they allow software to be written relatively easily. Register machines are dependent on sophisticated compilers; the simplicity of a register machine is more than outweighed by the complexity of the necessary compilers. Although the performance of a stack machine will often be inferior to that of a register machine, this is not as important as freedom and software reuse. A stack machine is the best compromise for the design goals of Freeputer. This has now been confirmed by extensive thought and experiments carried out over many weeks, although it was not the expected conclusion at the time. In other words, Freeputer 1.0 really was on the right track after all. All that is needed is refinement, not a change of direction.