End of hybrid experiments: about to return to stack machine.
Removed vR (replaced by cs).

CONCLUSION: Indeed, although it is a pain having to do such things as swap and rot and drop, on balance a stack machine is the better architecture for Freeputer since it results in a balanced distribution of complexity which is optimal, keeping the compiler relatively simple and keeping factoring relatively simple. Otherwise tedious pushing and popping of registers would get out of hand and would not be worth it. For a non-trivial application a stack machine is more maintainable for human beings despite slower performance at times. Probably the most significant discovery of all this experimentation is that a counter stack is really needed, as the lack of same results in undesirable clumsiness of stack-machine code. Having a counter stack makes coding simpler and easier and more concise.