Stack machine: starting from 'qmisc.c'.
Starting point is a copy of 'qmisc-native.c' so as to experiment with native Harvard implementation. This has not yet been converted from a register machine to a stack machine; that work is about to begin. Meanwhile the previous interpreted version of 'qmisc.c' has been moved to 'qmisc-interpreted.c' for future reference but it is a register machine.