Updated README to reflect chosen strategy for FVM 2.0.
After months of research and experimentation with various design philosophies, Plan A (with its generalized branch-on-failure strategy which is not particularly easy for programmers to use or understand) is now rejected as too complex and too dissimilar to the way human beings think about algorithms. What has certainly emerged is that a stack machine (rather than a register machine) is the best way forward and that in this respect FVM 1.0 was certainly on the right track. Register machines are for compilers. Stack machines are for humans. Thus what is needed is refinement not a complete change of tack. The move to Harvard architecture (away from the Von Neumann architecture of FVM 1.0) is justified as it will facilitate fast native implementations, upon which interpreted implementations can be self-virtualized with reasonable performance. The addition of a counter stack is an overdue improvement. The current README now reflects my latest thinking on the best way forward. It takes into account compromises derived from some limitations of array indexing in C and also some limitations of moving to FW32 encoding for interpreted implementations.