MILESTONE: Basic I/O strategy working (in, out).
After many months of thought, it has been decided that FVM2 will use a far simpler I/O strategy than FVM1. Specifically, there will only be two I/O instructions: in, out. They will branch on failure. They can move one single word to/from the data stack. There will always be only one output stream (for out) and one input stream (for in). Thus the FVM itself can be extremely simple. The responsibility for sophisticated I/O capability is instead exported to the environ in which the FVM is running. The environ can be thought of as a container or as an I/O device or as an I/O controller.

This simple JavaScript implementation uses the most rudimentary environ possible, lacking any control logic. However, there is no limit to the possible sophistication of an environ. In such a sophisticated environ, the FVM would issue commands to its environ (via out) by means of an explicit protocol; that is, a protocol not inherent in the FVM implementation but explicitly coded into programs running within the FVM. Such commands could have the effect of allowing the FVM to send/receive words to/from many destinations/sources. There is in principle no limit to this other than the degree of sophistication of the environ.

Of course, all this means that any program which performs I/O therefore requires that the FVM instance in which it is running must be deployed within a suitable environ; that is, an environ which fully and correctly and appropriately supports the explicit protocol used by that program to control its environ.

Furthermore, please note that there are no interrupts. This is a simple design decision which makes sense because the FVM is primarily intended to be used for algorithms and computations, rather than for real-time control systems. However, in theory any kind of system could be composed from multiple FVM instances deployed in appropriate environs.