Now using Web Worker.
The FVM instance now runs as a Web Worker; that is, effectively it runs in a different thread. This avoids freezing the browser during long runs. However, this implementation is incomplete in that stdin, stdout and stdtrc are not yet functioning correctly. For now, see the browser console for stdtrc.