I/O is now working correctly with Web Worker.
I/O is now working correctly for stdin, stdout and stdtrc. This was a non-trivial exercise due to the pain of getting vanilla JavaScript to do blocking input in the context of a virtual machine implemented as a Web Worker and without freezing the browser. The only reasonable solution at this point in time was to use SharedArrayBuffer and Atomics, thus unfortunately requiring features of ES2017. Therefore this solution will not run on older browsers. It is noteworthy that implementing this kind of virtual machine in C (or even in x86 assembly language on Linux) is far simpler and easier than implementing it in JavaScript! However, JavaScript offers portability (at the cost of relatively poor performance).

In future a port to Web Assembly will probably be done, to achieve far better performance than that currently offered by JavaScript.