Experiment: local vs global labels.
First stage of implementation in the assembler. Locals begin with s, globals begin with g, and it is possible to redefine locals. Allowing this redefinition means some preprocessing of the source file would need to be done before it could be compiled as C (but that would be possible using gcc local labels within a block). This initial implementation is crude and incomplete. Nevetheless, clearly this is a step forward and broadly in the right direction.