Basic m{ u{ x: u: s: }u }m scheme in place.
After considerable thought it was decided to keep the namespace scheme very simple and minimalistic. This means in practice that either: (i) a simple registry of some kind needs to be manually maintained so as to avoid the accidental reuse of module numbers; or (ii) a conversion tool sometimes needs to be used to renumber some modules to avoid namespace collisions when modules from different sources are combined into a single source. These drawbacks, although sometimes inconvenient, reduce complexity and are hardly onerous; they also make C compatibility more achievable.