Simplifying VM: radical experiment with generic stacks.
This radical experiment has just begun. The nature of the experiment is to move away from having a fixed number of stacks provided by the VM implementation and instead support the programmatic creation of as many generic stacks as the programmer desires. In this new model, stacks would reside in program memory, and thus be accessible via @ and ! instructions, just like any other part of program memory. Should, in a future experiment, the program counter (PC) also be moved into program memory, then the entire state of the VM would be contained in its accessible memory in a very neatly encapsulated manner.