I/O: further experimentation with in, out.
This initial, rudimentary, in-browser implementation of in and out is very slow and rendering is extremely slow. However, this will do for now until the design and implementation of the rest of the VM is finished, at which time schemes for browser environs with higher performance of in and out can be devised.