Implemented new RMW and ROM instructions.
This means that the instruction set planned for FVM 2.0 has now been, at least in some rudimentary manner, fully implemented in 'fvm2.js' although the behaviour of some instructions is not yet perfect in some corner cases and may yet be changed in some other cases. The soul of the design is now in place and can now meaningfully be experimented with. The JavaScript implementation is the only implementation which has reached this stage so far.