Alternative approach: 'qmisc.c' with link stack.
The proposed design for FVM 2.0, as outlined in 'pre-alpha2.0/README.md', is good but still requires an *annoying* amount of trapping and stack functionality to be coded in whatever implementation language you are using (such as assembly language, C, C++, Java, JavaScript or Web Assembly). It might indeed be that this is appropriate and proper and simply has to be accepted. However, perhaps the current 'README.md' is too pessimistic. Perhaps a virtual machine like 'qmisc.c' *could* be sufficiently practical if it had just one simple stack, of a small fixed size, whose only purpose was to hold return addresses. All other functionality could be composed programmatically in its instruction set and run natively due to the advantages of the Harvard Architecture. This would reduce complexity of VM implementation and allow greater flexibility so as to more often keep things simple.