Parking 'fvm2.js' as 'fvm2-mainline.js'.
About to start hacking 'fvm2.js' into a QMISC approach. If not happy with the results of same, later will roll back to the state of 'fvm2-mainline.js'. Essentially this is to make a final choice between the two strategies: QMISC (with a link stack, see current 'qmisc.c') or a 4-stack machine.