QMISC again rejected, this time finally.
Performance is inadequate as soon as one needs to emulate stacks with boundary checking. Once again, register machines are for compilers, stack machines are for humans. QMISC is more of a niche machine for simple programs not requiring stacks. The 4-stack-machine design currently described in 'pre-alpha2.0/README.md' is a much better overall compromise between performance and ease of programming.

This means that the *annoying* amount of implementation work for that 4-stack-machine (pertaining to trapping and stack functionality) simply has to be accepted as the *minimum necessary amount* of annnoying complexity which cannot practically be avoided.