Back to the 4-stack-machine design.
Work on Freeputer 2.0 shall now slowly proceed according to the design described in 'pre-alpha2.0/README.md' and accepting that it will take some time to gradually roll out the VM implementation in the various target languages including C, x86 assembly language, Java and JavaScript. Only relatively small refinements to that design are expected. The focus will be on C now, to sanity check the implementation, before returning to JavaScript and the other target languages.

The JavaScript implementation, although rudimentary and incomplete, has already progressed far enough to have proven the concept in general (see 'fvmui.html', 'fvm2.js', 'fvma.js' and 'prg.js') and to have demonstrated that such a virtual machine can run effectively in Firefox and Chrome. It is now necessary to shift the focus to C so that I/O performance can be sanity checked in a POSIX environment. Thereafter Emscripten might possibly be used to convert such a C implementation to auto-generated JavaScript (the nature and performance of which could then be compared to the existing JavaScript prototype) and to Web Assembly (which hopefully should yield major in-browser performance improvements). The C implementation will also be trialled on microcontrollers. The JavaScript implementation might then be ported to Node.js. The x86 and Java implementations are likely to be done last.

It is worth noting that the amount of effort required to implement this proposed FVM 2.0 design is perhaps not more than half the effort which was required to implement the FVM 1.0 design. That is, the redesign efforts have been worthwhile.