C implementation: adding NaN.
This is experimental and might be rolled back to trapping instead if the propagation of NaN proves to be impractical. However, it is a worth a try before making a final decision, since the use of NaN greatly simplifies programming in some respects, so long as its existence is always kept in mind.